optimization an anim fix
**Describe Pull request**
Removed the drawtext3d because the system already uses the qb-core drawtext function.
Also removed the marker for better optimization and lower resmon values.
There was no animation for picking up the garbage, set up a progressbar with animation for picking up the bags.

If your PR is to fix an issue mention that issue here

**Questions (please complete the following information):**
- Have you personally loaded this code into an updated qbcore project and checked all it's functionality? [yes] (Be honest)
- Does your code fit the style guidelines? [yes]
- Does your PR fit the contribution guidelines? [yes]